feat(MenuItem): adding "raw" prop (#721)
This pull request introduces a new feature for handling raw menu items
in the `Menu` component and updates the documentation and tests
accordingly. The most important changes include adding support for raw
menu items in `MenuItem`, creating a new story in the documentation, and
updating the test suite to cover the new functionality.

### New Feature: Raw Menu Items

*
[`packages/ui-menu/src/components/Menu/MenuItem.tsx`](diffhunk://#diff-b98fa67669fc6c42b3c8bca567d39ccb4b5fae4d2f6cf34345dbe5749b72f3efL11-R60):
Added support for raw menu items by introducing the `raw` and
`ignoreClick` props. Modified the component to render a `div` with
`role="menuitem"` when `raw` is true.
[[1]](diffhunk://#diff-b98fa67669fc6c42b3c8bca567d39ccb4b5fae4d2f6cf34345dbe5749b72f3efL11-R60)
[[2]](diffhunk://#diff-b98fa67669fc6c42b3c8bca567d39ccb4b5fae4d2f6cf34345dbe5749b72f3efL39-R73)

### Documentation Updates

*
[`packages/documentation/src/Components/Menu.stories.tsx`](diffhunk://#diff-38e03e431f106b2c80e5e6c37a827f0c600274e8fc7bb7280b1233c4440a8d23R169-R214):
Added a new story `RawItem` to demonstrate the usage of raw menu items
with `ToggleGroup` inside a `MenuItem`.
*
[`packages/documentation/src/Components/Menu.stories.tsx`](diffhunk://#diff-38e03e431f106b2c80e5e6c37a827f0c600274e8fc7bb7280b1233c4440a8d23R13):
Imported `ToggleGroup` and `ToggleGroupItem` from
`@versini/ui-togglegroup` to support the new story.

### Test Suite Enhancements

*
[`packages/ui-menu/src/components/Menu/__tests__/Menu.test.tsx`](diffhunk://#diff-eb0374cce0d1b1e35be93af5f58eb160b6cf419f1bef5a82a7759fa35034ca7fR34-R47):
Added new tests to verify the behavior of raw menu items, including
their rendering and click handling.
[[1]](diffhunk://#diff-eb0374cce0d1b1e35be93af5f58eb160b6cf419f1bef5a82a7759fa35034ca7fR34-R47)
[[2]](diffhunk://#diff-eb0374cce0d1b1e35be93af5f58eb160b6cf419f1bef5a82a7759fa35034ca7fR275-R314)