feat(Button): adding variant "selected" (#816)
This pull request introduces a new "selected" variant for buttons in the
`ui-button` package, along with the necessary styling and utility
updates. Additionally, it includes changes to the `ui-styles` package to
support the new variant with appropriate color tokens. Finally, a
development script is added to the `ui-types` package.

### Button Variant Addition:
*
[`packages/ui-button/src/components/Button/__tests__/Button.test.tsx`](diffhunk://#diff-bcbff46154f4ebc258bd11d085a90b403b6c47305f822485093615282d765af9R303-R366):
Added tests for rendering dark and light buttons with the "selected"
variant.
*
[`packages/ui-button/src/components/Button/utilities.ts`](diffhunk://#diff-dc1703275e22178b8358162e45b8752e3e08e0482680468f047707eca025d7d6R123-R136):
Updated utility functions to handle the "selected" variant for base,
hover, active, and border classes.
[[1]](diffhunk://#diff-dc1703275e22178b8358162e45b8752e3e08e0482680468f047707eca025d7d6R123-R136)
[[2]](diffhunk://#diff-dc1703275e22178b8358162e45b8752e3e08e0482680468f047707eca025d7d6R177-R186)
[[3]](diffhunk://#diff-dc1703275e22178b8358162e45b8752e3e08e0482680468f047707eca025d7d6R227-R236)
[[4]](diffhunk://#diff-dc1703275e22178b8358162e45b8752e3e08e0482680468f047707eca025d7d6R273-R282)

### Styling Support:
*
[`packages/ui-styles/src/plugins/tailwindcss/tokens.ts`](diffhunk://#diff-76e1ba53c5a19cf85819b9e79dcb6a86bae9ae0649e1cc7c91a26f7e17314362R26-R33):
Added color tokens for the "selected" variant, including dark, light,
and border colors.
[[1]](diffhunk://#diff-76e1ba53c5a19cf85819b9e79dcb6a86bae9ae0649e1cc7c91a26f7e17314362R26-R33)
[[2]](diffhunk://#diff-76e1ba53c5a19cf85819b9e79dcb6a86bae9ae0649e1cc7c91a26f7e17314362R90-R93)

### Development Script:
*
[`packages/ui-types/package.json`](diffhunk://#diff-c61e67b2f839cb9449cc6cd65947f877ac045ce6edfab2ca27c3f6fb45ff622bR23-R24):
Added a `dev:types` script for watching type changes during development.