feat(ButtonCopy): adding support for function for copyToClipboard prop (#842)
This pull request includes changes to enhance the functionality and
testing of the `ButtonCopy` component in the `ui-button` package. The
most important changes include adding support for a function as the
`copyToClipboard` prop, updating related stories, and adding new tests
to ensure the functionality works as expected.

Enhancements to `ButtonCopy` component:

*
[`packages/ui-button/src/components/Button/ButtonCopy.tsx`](diffhunk://#diff-24ea81118553e6719abd08feb8a415d9bf661dbffc0da007192f54ad8b4fa8bbL14-R20):
Added support for `copyToClipboard` prop to accept a function, which
allows dynamic text to be copied to the clipboard.

Updates to stories:

*
[`packages/api/src/stories/Components/ButtonCopy.stories.tsx`](diffhunk://#diff-eb56e67f4076cf8a641b63bd854b18fafda7caad6d3daad3908ed7e35b3d1187L25-R37):
Updated the `Basic` story and added a new `FunctionAsProp` story to
demonstrate the new functionality where `copyToClipboard` can be a
function.

Improvements to tests:

*
[`packages/ui-button/src/components/Button/__tests__/ButtonCopy.test.tsx`](diffhunk://#diff-fe4e5068a53104abebca0e5809a30dcbef4e366e7851fbeee0c38110bbf0c024L26-R26):
Updated existing test to specify it tests a string `copyToClipboard`
prop and added a new test to verify the functionality when
`copyToClipboard` is a function.
[[1]](diffhunk://#diff-fe4e5068a53104abebca0e5809a30dcbef4e366e7851fbeee0c38110bbf0c024L26-R26)
[[2]](diffhunk://#diff-fe4e5068a53104abebca0e5809a30dcbef4e366e7851fbeee0c38110bbf0c024R40-R54)